requirements.txt
python-docx
lightautoml
pymorphy2
scikit-learn
pytelegrambotapi
pymongo
tqdm
pandas
numpy
joblib

